Move Java classes under java top-level dir
Create per-directory BUILD files (not sure why this was not done before).

Add scripts to build and run Android Docker image.

Update CI to use new scripts.

Update Android Docker image to a later version of Debian and JDK, as
well as newer Android SDK and API.

Ref #2533
