+++ WORKSPACE
@@ -289,8 +289,8 @@ load("@build_bazel_rules_android//android:rules.bzl", "android_ndk_repository", 
 
 android_sdk_repository(
     name = "androidsdk",
-    api_level = 28,
-    build_tools_version = "28.0.3",
+    api_level = 30,
+    build_tools_version = "30.0.0",
 )
 
 android_ndk_repository(
