+++ android.Dockerfile
@@ -1,6 +1,7 @@
-# Download Debian-9 (Stretch) with openjdk-8 that is required for Android.
-ARG debian_snapshot=stretch-20200327
-FROM debian/snapshot:${debian_snapshot}
+# Use fixed snapshot of Debian to create a deterministic environment.
+# Snapshot tags can be found at https://hub.docker.com/r/debian/snapshot/tags?name=bullseye
+ARG debian_snapshot=sha256:d647c95797c43e8ef0e0667c4b4acba752ae70385dcf648877903f15c8977da4
+FROM debian/snapshot@${debian_snapshot}
 
 RUN apt-get --yes update && \
     apt-get install --no-install-recommends --yes \
@@ -15,7 +16,7 @@ RUN apt-get --yes update && \
     libmpfr-dev \
     libncurses5 \
     ocaml-nox \
-    openjdk-8-jdk \
+    openjdk-11-jdk \
     python-dev \
     python2.7-dev \
     python3-dev \
@@ -33,8 +34,8 @@ RUN apt-get --yes update && \
     && git --version
 
 # Use a fixed version of Bazel.
-ARG bazel_version=3.0.0
-ARG bazel_sha256=dfa79c10bbfa39cd778e1813a273fd3236beb495497baa046f26d393c58bdc35
+ARG bazel_version=5.3.1
+ARG bazel_sha256=1e939b50d90f68d30fa4f3c12dfdf31429b83ddd8076c622429854f64253c23d
 ARG bazel_url=https://storage.googleapis.com/bazel-apt/pool/jdk1.8/b/bazel/bazel_${bazel_version}_amd64.deb
 RUN curl --location "${bazel_url}" > bazel.deb \
     && sha256sum --binary bazel.deb && echo "${bazel_sha256} *bazel.deb" | sha256sum --check \
@@ -45,10 +46,11 @@ RUN curl --location "${bazel_url}" > bazel.deb \
 
 # Install Android SDK.
 # https://developer.android.com/studio/#downloads
-ARG android_sdk_version=4333796
+# https://developer.android.com/studio/index.html#command-tools
+ARG android_sdk_version=8512546
 ENV ANDROID_HOME /opt/android-sdk
 RUN mkdir --parents ${ANDROID_HOME} \
-    && curl --location https://dl.google.com/android/repository/sdk-tools-linux-"${android_sdk_version}".zip > android_sdk.zip \
+    && curl --location https://dl.google.com/android/repository/commandlinetools-linux-"${android_sdk_version}_latest".zip > android_sdk.zip \
     && unzip android_sdk.zip -d ${ANDROID_HOME} \
     && rm android_sdk.zip
 
@@ -56,12 +58,11 @@ RUN mkdir --parents ${ANDROID_HOME} \
 # https://developer.android.com/studio/releases/platform-tools
 # https://developer.android.com/studio/releases/platforms
 # https://developer.android.com/studio/releases/build-tools
-# '28.0.3' is the maximal version supported by NDK.
-ARG platform=28
-ARG tools=28.0.3
-RUN ${ANDROID_HOME}/tools/bin/sdkmanager --update \
-    && yes | ${ANDROID_HOME}/tools/bin/sdkmanager --licenses \
-    && yes | ${ANDROID_HOME}/tools/bin/sdkmanager \
+ARG platform=30
+ARG tools=30.0.0
+RUN ${ANDROID_HOME}/cmdline-tools/bin/sdkmanager --update --sdk_root=${ANDROID_HOME} \
+    && yes | ${ANDROID_HOME}/cmdline-tools/bin/sdkmanager --licenses --sdk_root=${ANDROID_HOME} \
+    && yes | ${ANDROID_HOME}/cmdline-tools/bin/sdkmanager --sdk_root=${ANDROID_HOME} \
     'tools' 'platform-tools' 'cmake;3.6.4111459' \
     "platforms;android-${platform}" "build-tools;${tools}" \
     "system-images;android-${platform};default;x86_64"
@@ -70,43 +71,12 @@ RUN ${ANDROID_HOME}/tools/bin/sdkmanager --update \
 ENV PATH "${PATH}:${ANDROID_HOME}/emulator:${ANDROID_HOME}/tools:${ANDROID_HOME}/platform-tools:${ANDROID_HOME}/tools/bin"
 ENV LD_LIBRARY_PATH "${LD_LIBRARY_PATH}:${ANDROID_HOME}/emulator/lib64:${ANDROID_HOME}/emulator/lib64/qt/lib"
 
-# Create an Android emulator.
-# `no` is used to opt-out from custom hardware profile.
-RUN echo no | ${ANDROID_HOME}/tools/bin/avdmanager create avd \
-    -n "android-${platform}-x86_64" \
-    -k "system-images;android-${platform};default;x86_64" \
-    -b x86_64
-
 # Install Android NDK
 # https://developer.android.com/ndk/downloads
-ARG android_ndk_version=r20b
+ARG android_ndk_version=r25b
 ENV ANDROID_NDK_HOME /opt/android-ndk
 RUN mkdir --parents ${ANDROID_NDK_HOME} \
-    && curl --location https://dl.google.com/android/repository/android-ndk-"${android_ndk_version}"-linux-x86_64.zip > android_ndk.zip \
+    && curl --location https://dl.google.com/android/repository/android-ndk-"${android_ndk_version}"-linux.zip > android_ndk.zip \
     && unzip android_ndk.zip -d ${ANDROID_NDK_HOME} \
     && mv ${ANDROID_NDK_HOME}/android-ndk-${android_ndk_version}/* ${ANDROID_NDK_HOME} \
     && rm --recursive --force android_ndk.zip
-
-# Bazel requires HOME variable to be set.
-# https://github.com/bazelbuild/bazel/issues/471
-ENV HOME /workspace
-
-# We use the `docker` user in order to maintain library paths on different
-# machines and to make Wasm modules reproducible.
-ARG USERNAME=docker
-
-# Placeholder args that are expected to be passed in at image build time.
-# See https://code.visualstudio.com/docs/remote/containers-advanced#_creating-a-nonroot-user
-ARG USER_UID=1000
-ARG USER_GID=${USER_UID}
-
-# Create the specified user and group.
-#
-# Ignore errors if the user or group already exist (it should only happen if the image is being
-# built as root, which happens on GCB).
-RUN (groupadd --gid=${USER_GID} ${USERNAME} || true) \
-    && (useradd --shell=/bin/bash --uid=${USER_UID} --gid=${USER_GID} --create-home ${USERNAME} || true)
-
-# Set the default user as the newly created one, so that any operations performed from within the
-# Docker container will appear as if performed by the outside user, instead of root.
-USER ${USER_UID}
