+++ java/src/main/java/com/google/oak/functions/android/client/AndroidManifest.xml
