+++ java/src/main/java/com/google/oak/functions/android/client/res/values/colors.xml
