+++ java/src/main/java/com/google/oak/remote_attestation/BUILD
@@ -1,5 +1,5 @@
 #
-# Copyright 2021 The Project Oak Authors
+# Copyright 2022 The Project Oak Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,7 +14,6 @@
 # limitations under the License.
 #
 
-load("@build_bazel_rules_android//android:rules.bzl", "android_library")
 load("@rules_java//java:defs.bzl", "java_library")
 
 package(
@@ -25,11 +24,11 @@ package(
 java_library(
     name = "remote_attestation",
     srcs = [
-        "src/com/google/oak/remote_attestation/AeadEncryptor.java",
-        "src/com/google/oak/remote_attestation/ClientHandshaker.java",
-        "src/com/google/oak/remote_attestation/KeyNegotiator.java",
-        "src/com/google/oak/remote_attestation/Message.java",
-        "src/com/google/oak/remote_attestation/SignatureVerifier.java",
+        "AeadEncryptor.java",
+        "ClientHandshaker.java",
+        "KeyNegotiator.java",
+        "Message.java",
+        "SignatureVerifier.java",
     ],
     deps = [
         "@com_google_guava_guava",
@@ -46,11 +45,11 @@ java_library(
 android_library(
     name = "remote_attestation_android",
     srcs = [
-        "src/com/google/oak/remote_attestation/AeadEncryptor.java",
-        "src/com/google/oak/remote_attestation/ClientHandshaker.java",
-        "src/com/google/oak/remote_attestation/KeyNegotiator.java",
-        "src/com/google/oak/remote_attestation/Message.java",
-        "src/com/google/oak/remote_attestation/SignatureVerifier.java",
+        "AeadEncryptor.java",
+        "ClientHandshaker.java",
+        "KeyNegotiator.java",
+        "Message.java",
+        "SignatureVerifier.java",
     ],
     # Exclude Android specific targets because they will not run in the main Docker Image.
     tags = ["manual"],
