+++ java/src/main/java/com/google/oak/remote_attestation/Message.java
