+++ java/src/test/java/com/google/oak/remote_attestation/MessageTest.java
