+++ scripts/android_docker_run
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+
+readonly SCRIPTS_DIR="$(dirname "$0")"
+# shellcheck source=scripts/common
+source "${SCRIPTS_DIR}/common"
+
+docker run "--volume=$PWD:/workspace" --workdir=/workspace gcr.io/oak-ci/oak-android:latest "$@"
