+++ scripts/build_examples_android
@@ -4,4 +4,14 @@ readonly SCRIPTS_DIR="$(dirname "$0")"
 # shellcheck source=scripts/common
 source "$SCRIPTS_DIR/common"
 
-"${SCRIPTS_DIR}/build_example_android" -e oak_functions
+bazel_build_flags+=(
+  '--fat_apk_cpu=x86_64'
+  '--linkopt=-llog'
+  '--symlink_prefix=bazel-client-android-'
+  '--host_crosstool_top=@bazel_tools//tools/cpp:toolchain'
+  # TODO(#3326): Remove this exception.
+  '--incompatible_java_common_parameters=false'
+)
+
+# Build everything, including targets marked as manual, which are normally excluded by the main Docker image.
+bazel build --build_manual_tests "${bazel_build_flags[@]}" //java/...:all
